mv syntax_errors to parser db
